Fixed bug; updated format for boolean check
Fixed bug where a segfault would occur if there was no 'x' in the
expression of a tree.
From now on I will write booleans with the Lvalue to the right of the
statement. This will prevent mistakes where the Lvalue is set to the
value by accident. ((0 == a) instead of a == 0))