Interpreter window is now interchanged between different assumptions
Multiple assumptions have been added to use the interpreter and the
interpreter window now updates its text according to the current interpreter
pointer.
Changed struct user_input so that it has "free" members that can be updated
if another variables should be added. This is done so that the ".dc" files
don't have their memory allocation changed so that older files can still be
read