Fixed segfaults
In the function update_user_interface the function would try and update
the interface even if the values had not been set yet in the hashtable.
This should only be done once values have been set.
In the function replant there were two bug, once in the assert which tested
for UI_ADMINCOST instead of UI_AMOUNT and another when the tree is planted
we need to strclean it first.