fixed a bug regarding project
variables were set to be projected with inflation, but at the time
inflation amount was determined, it was set to zero. afterwards this
was no longer updated as the interpreter had already been created with
zero as the amount to project. the interpreters had to be reset after
the interface checking.